chg: refactored export subsystem
