+++ c3shop/frontpage/views.py
@@ -1,4 +1,6 @@
 from django.http import HttpResponse, HttpResponseForbidden
+
+from frontpage.management.export import export_dbhints, export_statistics, export_invoice
 from .uitools.footerfunctions import render_footer
 from .uitools.headerfunctions import render_content_header
 from .uitools.body import *
@@ -7,7 +9,7 @@ from .management import random_actions
 from frontpage.management.mediatools import media_page, media_upload_page, media_actions, media_select
 from .management import edit_reservation, reservation_processing, settings_page
 from frontpage.management.articletools import article_select, article_actions, article_page, edit_article
-from .management import edit_settings, export, password_page
+from .management import edit_settings, password_page
 from frontpage.management.grouptools import edit_group
 
 from .management.grouptools.grouparticlesupdate import handle_group_articles_request
@@ -136,11 +138,13 @@ def admin_list_users(request):
 def action_save_post(request):
     return edit_post.do_edit_action(request, "/admin/posts")
 
+
 def action_change_password(request: HttpRequest):
     response = require_login(request, min_required_user_rights=0)
     if response:
         return response
     return password_page.action_change_password(request)
+
 
 def action_save_user(request):
     return edit_user.action_save_user(request, "/admin/users")
@@ -206,11 +210,11 @@ def admin_export(request: HttpRequest):
         return response
     if request.GET.get("method"):
         if request.GET["method"] == "pdf":
-            return export.export_orders_to_pdf(request, request.GET["reservations"].split(','))
+            return export_invoice.export_orders_to_pdf(request, request.GET["reservations"].split(','))
         if request.GET["method"] == "rejectstatistics":
-            return export.export_reject_statistics(request)
+            return export_statistics.export_reject_statistics(request)
         if request.GET["method"] == "datadump":
-            return export.request_data_dump(request)
+            return export_dbhints.request_data_dump(request)
     return HttpResponseForbidden()
